fix(security)!: unsafe org-confirm-babel-evaluate set to nil
The org-confirm-babel-evaluate is an unsafe variable to set to nil
as a global value, especially if the mode sets it without any
warning or notice.

This commit removes this unsafe setup as a security measure. Now the
users can, if they wish so, set this value in their init file, or add
a hook to notebook-mode, or (most secure way to do it) set it as a
file-local variable.

The notebook.org document implements the last solution as an example.

BREAKING CHANGE: org-confirm-babel-evaluate setup now left to user.
Signed-off-by: Alexandre Martos <contact@amartos.fr>